feat(WebpackSSR): Async component loading for server & build output in dist directory
