+++ src/lib/webpack/dummy-loader.js
@@ -1,0 +1,3 @@
+module.exports = function(source, map) {
+  this.callback(null, source, map);
+};
